Add Dockerfile for running acceptance tests in containers
The setup is that we will have a container that:

 - has sshd running, and is configured to allow easy ssh to itself
 - has a ruby environment to run `bd` (the host will act as the
   initiating node)
 - has `git` and `rsync`, which would be run via ssh or sh
   depending on if this is remote or local deployment, thus
   this same host acts as the target node
 - we can either ssh in to run `bd`, or we can directly run
   via docker commands (it really doesn't matter)

Locally we will run the suite, which will spin up the container
and run specs against it. In CI, the CI system will spin up the
container and run specs against it.

Note: we can probably improve this Dockerfile, and almost certainly
can move away from using the root user in the container, but this
should be a workable start.